feat: move user-webauthn types and add auth-metadata
