+++ src/libs/satellite/src/user/core/assert.rs
@@ -1,13 +1,12 @@
 use crate::errors::user::{
-    JUNO_DATASTORE_ERROR_USER_CALLER_KEY, JUNO_DATASTORE_ERROR_USER_CALLER_WEBAUTHN_KEY,
-    JUNO_DATASTORE_ERROR_USER_CANNOT_UPDATE, JUNO_DATASTORE_ERROR_USER_INVALID_DATA,
-    JUNO_DATASTORE_ERROR_USER_INVALID_WEBAUTHN_DATA, JUNO_DATASTORE_ERROR_USER_KEY_NO_PRINCIPAL,
-    JUNO_DATASTORE_ERROR_USER_NOT_ALLOWED, JUNO_DATASTORE_ERROR_USER_WEBAUTHN_CANNOT_UPDATE,
+    JUNO_DATASTORE_ERROR_USER_CALLER_KEY, JUNO_DATASTORE_ERROR_USER_CANNOT_UPDATE,
+    JUNO_DATASTORE_ERROR_USER_INVALID_DATA, JUNO_DATASTORE_ERROR_USER_KEY_NO_PRINCIPAL,
+    JUNO_DATASTORE_ERROR_USER_NOT_ALLOWED,
 };
-use crate::user::core::types::state::{BannedReason, UserData, UserWebAuthnData};
+use crate::user::core::types::state::{BannedReason, UserData};
 use crate::{get_doc_store, Doc, SetDoc};
 use candid::Principal;
-use junobuild_collections::constants::db::{COLLECTION_USER_KEY, COLLECTION_USER_WEBAUTHN_KEY};
+use junobuild_collections::constants::db::COLLECTION_USER_KEY;
 use junobuild_collections::types::core::CollectionKey;
 use junobuild_shared::controllers::controller_can_write;
 use junobuild_shared::ic::api::id;
@@ -55,28 +54,7 @@ pub fn assert_user_collection_data(collection: &CollectionKey, doc: &SetDoc) -> 
     decode_doc_data::<UserData>(&doc.data)
         .map_err(|err| format!("{JUNO_DATASTORE_ERROR_USER_INVALID_DATA}: {err}"))?;
 
-    Ok(())
-}
-
-pub fn assert_user_webauthn_data(
-    caller: Principal,
-    collection: &CollectionKey,
-    doc: &SetDoc,
-) -> Result<(), String> {
-    let user_webauthn_collection = COLLECTION_USER_WEBAUTHN_KEY;
-
-    if collection != user_webauthn_collection {
-        return Ok(());
-    }
-
-    let data = decode_doc_data::<UserWebAuthnData>(&doc.data)
-        .map_err(|err| format!("{JUNO_DATASTORE_ERROR_USER_INVALID_WEBAUTHN_DATA}: {err}"))?;
-
-    let public_key = Principal::self_authenticating(&data.public_key.value);
-
-    if principal_not_equal(caller, public_key) {
-        return Err(JUNO_DATASTORE_ERROR_USER_CALLER_WEBAUTHN_KEY.to_string());
-    }
+    // TODO: assert user.auth_metadata is set if webauthn
 
     Ok(())
 }
@@ -104,25 +82,6 @@ pub fn assert_user_write_permission(
 
     // The user already exist but the caller is not a controller
     Err(JUNO_DATASTORE_ERROR_USER_CANNOT_UPDATE.to_string())
-}
-
-pub fn assert_user_webauthn_write_permission(
-    collection: &CollectionKey,
-    current_doc: &Option<Doc>,
-) -> Result<(), String> {
-    let user_webauthn_collection = COLLECTION_USER_WEBAUTHN_KEY;
-
-    if collection != user_webauthn_collection {
-        return Ok(());
-    }
-
-    // It's a new document
-    if current_doc.is_none() {
-        return Ok(());
-    }
-
-    // The user webauthn entry already exist
-    Err(JUNO_DATASTORE_ERROR_USER_WEBAUTHN_CANNOT_UPDATE.to_string())
 }
 
 pub fn assert_user_is_not_banned(
