+++ src/libs/satellite/src/user/mod.rs
@@ -1,3 +1,4 @@
 pub mod core;
 pub mod internal_hooks;
 pub mod usage;
+pub mod webauthn;
