+++ src/libs/satellite/src/user/webauthn/mod.rs
@@ -1,0 +1,5 @@
+pub mod assert;
+mod store;
+mod types;
+
+pub use store::delete_user_webauthn;
